refactor background into a new module
